requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
dateparser==1.1.8
plyer==2.1.0
python-dotenv==1.0.0